Conflicts:
	vpbuild/vppackage.go
	vpbuild/vpversion.go

+++ .utils/get.sh
@@ -19,12 +19,20 @@
 # Vapor homepage: http://www.ufoot.org/liquidwar/v7/vapor
 # Contact author: ufoot@ufoot.org
 
+if [ -d ../.utils ] ; then
+    cd ..
+fi
+if [ ! -d .utils ] ; then
+    echo "$0 should be run in srcdir"
+    exit 1
+fi
+
 export GOPATH=$(pwd)
 
-go get -u golang.org/x/crypto/ripemd160
-go get -u golang.org/x/crypto/openpgp
-go get -u golang.org/x/crypto/openpgp/packet
-go get -u github.com/golang/lint/golint
+go get golang.org/x/crypto/ripemd160
+go get golang.org/x/crypto/openpgp
+go get golang.org/x/crypto/openpgp/packet
+go get github.com/golang/lint/golint
 
 rm -rf src/github.com/ufoot/vapor
 install -d src/github.com/ufoot/vapor
